[TASK] Raise ext:styleguide extension versions
Raise ext:styleguide extension versions to
* 9.0.1
* 8.0.14

The styleguide extensions uses and invalid configuration for FAL
references using the table sys_file_reference as MM table - the
regular use case is to use that as 1:n foreign_field composite
assiciation. In a result the example in the styleguide extension
mixed the meaning and behavior of both fields uid_local and
uid_foreign.

Resolves: #83355
Releases: master, 8.7
Change-Id: I203b468b4a06e3196a98e2c5de1ba06419a65636
Reviewed-on: https://review.typo3.org/55114
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>