[TASK] CGL fix core after php-cs-fixer 2.9.0 raise
The "self_accessor" fixer is more strict with younger
php-cs-fixer releases and changes method argument type
hints to self if a class object of the same class is
expected. It also changes return type hints to self
in this case.
This fixer now may collide with the "xclass" system
of the core and is thus deactivated. We shoud decide
manually if self or class name is used.

Change-Id: I3a8c20bea1d79ac2c7bc1e9ecbc2d435c4047843
Resolves: #83360
Related: #83331
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/55121
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>