[TASK] v7 Changelog: Fix compile warnings, update 7.6.x/ Index.rst
Fixes a broken link and restructures the 7.6.x/Index.rst to be
free from patch level releases. This is ok with the low number
of files in this folder. Since v8, the install tool documentation
module will further help with keeping track of new files during
patch level releases, too. Not having named patch level versions
is easier to maintain from core point of view and simplifies
the underlying workflow.
The patch also adds indirectly adds a file which has been forgotton
to reference in Index.rst, this can't happen anymore from now on
due to the glob reference.

Change-Id: I2163f0823becbf5b719dc4cf9c1520ed224abb15
Resolves: #83361
Releases: 7.6
Reviewed-on: https://review.typo3.org/55124
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>