[TASK] impexp: Remove size handling
ext:impexp has a "size" display and adds size information
to export formats for records file sizes. This is entirely
removed with the patch - both from interface and .xml/.t3d
exports:

* The size handling on db records always made little sense,
  the default maximum size of 1MB was "never" hit, there
  wasn't even an input box to raise that limit, ever.
  This information is now just removed from the export formats.
* The size handling on files is rather irrelevant with FAL,
  since FAL file integrity is checked with sha1 already. The
  size display during export in the interface for FAL files
  never worked since FAL migration in core v6.0 - It always
  displayed 0. Nobody ever cared, there is not even a forge
  issue about that.
* Removing the 'size' column in the export interface - which
  only carried rather irrelevant information on record "size"
  only - simplifies the interface a bit.
* The record size information in exports depends on used
  dbms due to different internal database types. Removing that
  makes exports more dbms agnostic. This can be seen with
  all the functional test fixture file changes coming with
  the patch - removing that reduces a maintenance nightmare
  in this area and improves the chance to align exports
  in between the functional testes on different dbms again.

Note that exporting .xml/.t3d in v9 and importing it
in a v8 instance does not work anyway due to more
significant DB field related changes in v9 already.
The effects of this patch could be easily mitigated
on v8, but we were unable to create anything in v9 that
is able to be imported in v8 again - and ext:impexp
is not created to be an upgrade / download helper
in the first place. Thus, ignoring size checks in
v8 to mitigate effects of the file format change
does not seem to be necessary and is not done.

Change-Id: I0cae7b371b6ac027e27c9192ffa273dabc9795c7
Resolves: #83606
Releases: master
Reviewed-on: https://review.typo3.org/55391
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Reiner Teubner <rteubner@me.com>
Tested-by: Reiner Teubner <rteubner@me.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>