[FEATURE] Add new main module "Site"
A new main module for the TYPO3 Backend is introduced, named "site",
which serves a main entry point for integrators and site admins to configure
the "Frontend" and the "How" of a TYPO3 website, and not the "what" (= content).

The site management module's purpose for now is to include Redirect Management,
but will be extended during v9 development to also contain Domain/Site/404
URL handling, TypoScript and Templating / Fluid configuration.

Resolves: #83637
Releases: master
Change-Id: I06c88d09fa16a40fa8c0f6176ed05f002629498c
Reviewed-on: https://review.typo3.org/55425
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Reiner Teubner <reiner.teubner@me.com>
Tested-by: Reiner Teubner <reiner.teubner@me.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>