[TASK] Ensure HTTP RequestHandlers always return a PSR-7 Repsonse
This is in preparation for PSR-15 middleware support which
will require PSR-7 RespnseInterface return type declarations
for request handlers.

As TSFE powers a concept of outputting nothing [see isOutputting()]
we need to add NullResponse (which implements the PSR-7 ResponseInterface)
which Core/Bootstrap can detect, to stop invoking header() and echo.

Change-Id: Ie3169a4365a85d0472523138cc73bb47cbbcb70f
Releases: master
Resolves: #83724
Reviewed-on: https://review.typo3.org/55498
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>