[BUGFIX] Refactor record querying in deep nested structures in recycler
This patch refactors how the recycler queries records in deep page
structures.
Instead of resolving each level of the page tree on demand and executing
the complex logic of creating query builders and paging all over again,
a list of page ids is generated once and stored in the Caching Framework
for the current request. This list is then used in an multiple `IN()`
statements (depending on the DBMS engine used).

Also, some wrong type hints and a wrong language file reference are
fixed.

Resolves: #83702
Releases: master, 8.7
Change-Id: Ie3314b5a9209fb5585f95a70c16b35639951c197
Reviewed-on: https://review.typo3.org/55472
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>