[BUGFIX] Deprecate broken buildQueryParameters hook
This change deprecates the "buildQueryParameters" hook in three
different classes and adds two new hooks in DatabaseRecordList
and PageLayoutView to modify the current query.

With #82334 a cleanup of AbstractRecordList has introduced
the same hook into multiple classes, which breaks existing
hooks, because of the sixth parameter which can be one of three
different classes without a common parent class or interface.
This makes it impossible to use a type hint in the hook class.

Another problem is: an extension which implements the hook
for the list module and uses a type hint will break the page module.

The same query manipulation can be achieved with the two
new hooks, which have a separate identifier.

Resolves: #83740
Related: #82334
Releases: master
Change-Id: Ie3b2c8082f86c6632400a8194dca4ca244b428bc
Reviewed-on: https://review.typo3.org/55512
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Susanne Moog <susanne.moog@typo3.org>