[TASK] Remove overkill extension classes in about module
When the about module (EXT:about) was refactored into
Extbase, a Repository and Domain Model for Extensions was
added to render title, key and authors of an extension.

For modules this was done directly in the controller,
which is 100% sufficient in this one-time case, instead
of doing it the Extbase(tm) persistence way.

This way, all data is fetched the same way in the about module.

The two domain model / repository classes are removed,
and due to the fact that they are not part of the API,
a RST file is not added.

Resolves: #83890
Releases: master
Change-Id: I1169336452d7497ea0ed2e1a99553c164ec78612
Reviewed-on: https://review.typo3.org/55710
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>