[FEATURE] Allow deactivation of FormDataProviders
To manipulate the FormEngine data provider list by extensions,
it can be helpful to just set a data provider as disabled
and add an own one after the disabled one and before the
next one. This avoids funny array munging and dependency
shuffling if an extension author needs to substitute an
entire data provider with an own solution.

The concept has been implemented for node expansion
render types in a similar way and is repeated here.

Change-Id: Ia1f2de94940248b60e851c7b9968539f5ab0ef79
Resolves: #83906
Releases: master
Reviewed-on: https://review.typo3.org/55724
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>