[TASK] Move FE OutputCompression to PSR-15
Output Compression should be separated from the request handling, and is
now moved into a PSR-15 middleware.

This change also decouples Output Compression from Bootstrap, and the
Request Handler, so it can be re-used in other areas.

Intentionally omitted is a proper cleanup (ob_get_clean) and an explicit
write to the response object (in the middleware). That's up for later
patches. The idea of this patch is to keep functionality identical for
now.

Resolves: #83931
Releases: master
Change-Id: Ic84707cac6c858698f290069f6aa492724ce0602
Reviewed-on: https://review.typo3.org/55746
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>