[TASK] Do not repopulate $GLOBALS['TYPO3_CONF_VARS'] in installer
This is a followup for Icc8a3482edaef1ea329e68638d5ef467548062fc
which missed to remove the TYPO3_CONF_VARS repopulation
from the installer controller.

Now that ExtensionConfiguration::set() updates changed properties,
the manual repopulation after calling
synchronizeExtConfTemplateWithLocalConfigurationOfAllExtensions()
is not needed anymore and should be removed for the same
reasons as described in the mentioned change.

Change-Id: I3ff24411d2f0807a09c49d19bc021e9a239e6cb0
Releases: master
Resolves: #83954
Related: #83958
Reviewed-on: https://review.typo3.org/55778
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>