[BUGFIX] Move file extension detection to FAL driver
The only object that is allowed to handle the physical file in the FAL
is the driver. As that's the only instance that knows how to access the
file.

This patch makes sure the Indexer and FileExtensionFilter asks the
driver of the file extension instead of using the PathUtility itself.

Releases: master, 8.7
Resolves: #83976
Change-Id: I80bec3c3f86d1137747e37dd99973ada8909ad36
Reviewed-on: https://review.typo3.org/56696
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>