[TASK] Remove additional GET variables for sys_preview links
The possibility of EXT:workspaces to add additional GET variables to
preview links was dropped with TYPO3 6.0, instead linking to the full
workspace was enabled, also of using workspaces instead of versioning.

This info (a string of additional GET vars) is now removed from adding
through the workspaces Hook class, instead the link to the whole
workspace is generated, and identified via a ADMCMD md5-hash.

Resolves: #84006
Releases: master
Change-Id: Ic62b106bb0d350214daf2deda9eae739a3ab9633
Reviewed-on: https://review.typo3.org/55857
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>