[TASK] Consolidate hooks in workspace preview
The Workspace preview hooks first hook into TSFE->connectToDB
to see if ADMCMD_prev is set via GP/cookie. If so, the TSFE object gets
re-instantiated, which is utterly unnecessary.

Additionally, it unsets the BE user cookie, solely to use a SECOND hook
(initializePreviewUser()) to check if the BE_USER is null.

Instead, the hooks are consolidated and the one hook after a regular BE user
exists, gets used, and overridden by a custom BE user object as defined
in the preview configuration with the workspace preview.

Resolves: #84014
Releases: master
Change-Id: Id3caf9cd99ff13c906062ff95523f6066d90d322
Reviewed-on: https://review.typo3.org/55866
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>