[TASK] Move Page ID resolving to PSR-15 middleware
The next PSR-15 middleware is introduced that was previously part of the
Frontend Request Handler. It defines the actual (current) implementation
to determine the Page ID and GET/POST parameters handled in TSFE.

A possible alternative could be a Router doing the same process at this
very position currently.

Resolves: #84096
Releases: master
Change-Id: I6aa9b5f35183034a8ebb93258d085acafb93862a
Reviewed-on: https://review.typo3.org/55961
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>