[TASK] Set COMPOSER_ROOT_VERSION to 9.2.0@dev for bamboo
If typo3/testing-framework needs to be synced with changes
in core it needs to require the dev branch of the upcoming
release as the compatibility to previous releases needs to
be dropped.
Therefore it has to define a dependency in composer.json like:
  "typo3/cms-core": "^9.2@dev"

If composer install is executed with COMPOSER_ROOT_VERSION
set to 9.1.0 that produces a conflict, therefore we raise
the version to 9.2.0@dev. That actually aligns the setting
with the branch alias (9.2.x-dev) set in composer.json.

Releases: master
Resolves: #84102
Change-Id: Ib9a6f1027cd738bcc1ecf00ef4e059ad179fc7c0
Reviewed-on: https://review.typo3.org/55965
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>