[BUGFIX] Harden deprecation log handling
Because TYPO3's DefaultConfiguration is recursively
merged with settings from LocalConfiguration, it was
impossible to change or remove log writers that are
defined in DefaultConfiguration. One would have had
to use AdditionalConfiguration for that.

A simpler way to disable individual log writers is added now.
It is simpler, because it allows to disable log writers
directly in LocalConfiguration.

To avoid misconfiguration regarding deprecation logging,
E_USER_DEPRECATED is now enforced in SYS/errorHandlerErrors.

E_USER_DEPRECATED needs to be enforced to avoid,
that the default PHP error handler is called, which would cause
these errors to be added to the output or filling up log files,
depending on the global PHP configuration (php.ini).

Because all other values from DefaultConfiguration.php resemble
production settings and logging deprecation messages in production
are not desired, the deprecation logger is now disabled by default.

Last but not least, deprecation logging is now enabled/disabled in
the Live and Debug presets in the install tool UI.

Resolves: #84105
Resolves: #89934
Resolves: #88444
Releases: master, 9.5
Change-Id: I52a6f9c70ad13e6e0bad6a6b06b6fbfe7abc623c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/56077
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>