[TASK] Merge the package DependencyResolver into the PackageManager
The DependencyResolver and the PackageManager have a cyclic dependency to
each other which is currently resolved using GeneralUtility::makeInstance.
As the DependencyResolver is actually only used for the
PackageManager – and relies on it – it can be merged, saving a lot of
hassle.

The DependencyResolver class is not marked @internal and is therefore
deprecated.

Releases: master
Resolves: #84109
Change-Id: I71adccec3f13eb6de859f065937fbcde369758fe
Reviewed-on: https://review.typo3.org/55977
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>