[BUGFIX] Change getUrl() $report['content_type'] to string
In the conversion of \TYPO3\CMS\Core\Utility\GeneralUtility::getUrl()
from cURL to GuzzleHttp the type of the $report['content_type'] field
was changed from a string to an array of strings.

This changes the field back to a string.

Also updated ErrorController getUrl test double to reflect the changes.

Resolves: #84173
Related: #84167
Releases: master, 8.7
Change-Id: Ia0eb60c42e920d66ce63afd0bd659c4b56d6d05b
Reviewed-on: https://review.typo3.org/56057
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>