[BUGFIX] Do not mark "current" links as broken
The linkhandler can resolve pages by uid or by using the
special keyword "current" which links the current page id.

The RteHtmlParser marked links to "current" as broken as the check
on whether the page record exists fails with a non-int uid.

The check has been disabled for links to the current page, as in that
context "current" has to be a valid page record (else content could
not be edited on it).

Resolves: #83423
Releases: master, 8.7
Change-Id: Iefde294a5f84e28a5faa1bc1262b7cc84e625f7e
Reviewed-on: https://review.typo3.org/56072
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>