[FEATURE] Add column selector to filelist
This adds the column selector component,
introduced in #94218, to the filelist module.

As known from recordlist, it can be used to
manage the fields to be displayed for each
file / folder. Fields to be selected are the
combination of special fields (e.g. read/write
permissions), sys_file fields and sys_file_metadata
fields.

Resolves: #84184
Releases: master
Change-Id: Ie1369e616f35dc4a92e90409eba202aa4afe15f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/56073
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>