[TASK] Move IRRE actions to a map and export inline
The possible requests done in jsfunc.inline.js are now stored in a map
that holds the method (e.g. "create") and points to a specific route for
AJAX requests.
Additionally, the inline object is made public for the time being and
explicitly marked as internal.

This allows overriding specific routes, which is required for the
upcoming "Sites" module.

Resolves: #84199
Releases: master
Change-Id: Ief9fc1382e60a6b894e2970f9349f9b03913d021
Reviewed-on: https://review.typo3.org/56098
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>