[BUGFIX] Trigger submit of RSA encrypted form properly
The change of #76120 changed the selector of possible submit buttons.
However, in Firefox a submit is not triggered when a focussed submit
element is clicked via JavaScript.

The code is changed to enforce a submit of the form instead.

RsaEncryptionWithLib.min.js created with:
../../../../../../Build/node_modules/uglify-js/bin/uglifyjs \
RsaLibrary.js RsaEncryption.js > RsaEncryptionWithLib.min.js

Resolves: #84253
Related: #76120
Releases: master, 8.7, 7.6
Change-Id: Id30f4536dccec28edc0d909114c8fde71614bc30
Reviewed-on: https://review.typo3.org/56142
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>