[BUGFIX] Add correct context to page tree context menu
In order to allow user to disable context menu items for certain context,
e.g. page tree, page tree need to correctly pass the context to the JS.

As SvgTree doesnt support context menu, the code is moved to the PageTree.js
Also a bug in the changelog documentation is fixed.

Releases: master, 8.7
Resolves: #84017
Change-Id: I4c3b51aa50d1750b86391342ba45aaec00c70a10
Reviewed-on: https://review.typo3.org/56342
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Łukasz Uznański <l.uznanski@macopedia.pl>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Paweł Rogowicz <rogowicz.pawel@gmail.com>
Tested-by: Paweł Rogowicz <rogowicz.pawel@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>