Revert "[BUGFIX] Test ifBlank with strlen again"
This reverts commit c12fe3c17c3d339e9c84e8867ac9f6433e4953ca.

The patch for issue #84434 introduced a different behavior for the
stdWrap.ifBlank feature. The current implementation now ignores
whitespace characters (tab, spaces, new lines, carriage return)
which was not the case in previous TYPO3 versions - as a comparison:

* TYPO3 v7: trim($content) === ''
* TYPO3 v6.2: !strlen(trim($content))
* TYPO3 v.4.5: !strlen(trim($content))
* TYPO3 v4.3: !strlen(trim($content)

Thus, the trim() part is basically missing currently. To avoid this
regression concerning the change of behavior, the change is reverted.

Related: #84434
Resolves: #84520
Releases: master, 8.7
Change-Id: I1fcbd0d249e18d06a96d481bf44a39de0662bed1
Reviewed-on: https://review.typo3.org/56422
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>