Update MediaViewHelper.php
Allow additionalAttributes in function render
In some cases, it is useful to call the Render-Fuction with additionalAttributes from your fluid-Template (e.g. by providing a poster-Attribut with a video-Tag like <f:media additionalAttributes="{poster: 'fallback-img.jpg'}" ) 
Alltough this additionalAttributes ist already registered by the Core at this position, the function render() did not made use of it so far.