[BUGFIX] Only apply language constraints for "Show hidden" checkbox if neccessary
The checkbox "Show hidden content elements" currently checks any selected
language. This is malfunctioning in case the language "Default" is
selected in the "Language" view, as it won't affect any translation.

The patch now changes the handling to add the language constraint only if
the view renders multiple languages and any language is selected, that is
not the default language.

Resolves: #84589
Resolves: #84573
Releases: master, 8.7
Change-Id: I572d594e5828d423e67a40481f44857df79ebae9
Reviewed-on: https://review.typo3.org/56590
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>