[TASK] Move linkData() functionality in PageLinkBuilder
The "mother of building all links in Frontend" is moved
into PageLinkBuilder. Along with this, building a
mountPointMap is moved into the PageLinkBuilder as well,
as this is solely needed.

Existing methods are deprecated:
 * TemplateService->linkData
 * TemplateService->getFromMPmap
 * TemplateService->initMPmap_create

The existing hook is moved into PageLinkBuilder as well,
so functionality like RealURL will work the same.

Resolves: #84637
Releases: master
Change-Id: I513bd721a77ed8959f3c8b4bfc4eaa25dd05eacc
Reviewed-on: https://review.typo3.org/56576
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>