[FEATURE] Introduce fluid data-processor for language menus
The HMENU is extended to support the auto filling of the special.value
with all languages defined for the current site. To each menu item the
corresponding SiteLanguage is appended as array to be available at
TypoScript.

This menu processor utilizes HMENU to generate a json encoded language
menu string that will be decoded again and assigned to FLUIDTEMPLATE as
variable.

Resolves: #84650
Resolves: #84775
Releases: master
Depends: #Iabeeb6835a98c8f5a71d502379ed63a68dfad6dd
Change-Id: I5b602256962deb47e89fc190401dc0281dc5ebb0
Reviewed-on: https://review.typo3.org/56597
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>