[BUGFIX] Hide save and new button in connected translation mode
In connected translation mode it should not be possible to use the
save and new button as this would lead to an inconsistent translation
state. This change adds an additional condition for tt_content records
to determine if the content is in free translation mode. Only in free
translation mode the save and new is shown.

Resolves: #84752
Releases: master, 8.7
Change-Id: Id500098ef9f642472e96cce59205b4a25377e421
Reviewed-on: https://review.typo3.org/56817
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>