[BUGFIX] Always exclude E_USER_DEPRECATED from exceptionalErrors
There is a limited use case for exceptions to be thrown
on deprecation messages. To ease the upgrade path to TYPO3 9
and to not allow "accidental" configuration that breaks the system,
deprecation messages will never throw exceptions.

Users with other use cases, can still implement and configure their
own exception handler.

Resolves: #84802
Releases: master
Change-Id: Iebbc2c201ba8bb68fa5725b90fba516d37d168be
Reviewed-on: https://review.typo3.org/56752
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jan Stockfisch <jan.stockfisch@googlemail.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>